backend/app/auth.py
from fastapi import Depends, HTTPException, status
from fastapi.security import HTTPBearer
from jose import JWTError, jwt
import os
from datetime import datetime, timedelta
from typing import Optional

security = HTTPBearer()

async def get_current_user(token: str = Depends(security)):
    credentials_exception = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="Could not validate credentials",
        headers={"WWW-Authenticate": "Bearer"},
    )
    try:
        # Return the token directly since it's the Spotify access token
        return {"access_token": token.credentials}
    except Exception:
        raise credentials_exception
