backend/requirements.txt
fastapi
python-dotenv
uvicorn
httpx
python-jose[cryptography]
python-multipart
